Release/1.1.0

+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/documentserver/managers/callback/DefaultCallbackManager.java
@@ -37,6 +37,7 @@ import org.springframework.stereotype.Component;
 import java.io.*;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
@@ -68,36 +69,37 @@ public class DefaultCallbackManager implements CallbackManager {
     @Autowired
     private ServiceConverter serviceConverter;
 
+    // save file information from the URL to the file specified
     private void downloadToFile(String url, Path path) throws Exception {
-        if (url == null || url.isEmpty()) throw new RuntimeException("Url argument is not specified");
-        if (path == null) throw new RuntimeException("Path argument is not specified");
+        if (url == null || url.isEmpty()) throw new RuntimeException("Url argument is not specified");  // URL isn't specified
+        if (path == null) throw new RuntimeException("Path argument is not specified");  // file isn't specified
 
         URL uri = new URL(url);
         java.net.HttpURLConnection connection = (java.net.HttpURLConnection) uri.openConnection();
-        InputStream stream = connection.getInputStream();
+        InputStream stream = connection.getInputStream();  // get input stream of the file information from the URL
 
         if (stream == null) {
             connection.disconnect();
             throw new RuntimeException("Input stream is null");
         }
 
-        storageMutator.createFile(path, stream);
+        storageMutator.createOrUpdateFile(path, stream);  // update a file or create a new one
     }
 
     @SneakyThrows
-    public void processSave(Track body, String fileName) {
+    public void processSave(Track body, String fileName) {  // file saving process
         String downloadUri = body.getUrl();
         String changesUri = body.getChangesurl();
         String key = body.getKey();
         String newFileName = fileName;
 
-        String curExt = fileUtility.getFileExtension(fileName);
-        String downloadExt = fileUtility.getFileExtension(downloadUri);
+        String curExt = fileUtility.getFileExtension(fileName);  // get current file extension
+        String downloadExt = fileUtility.getFileExtension(downloadUri);  // get an extension of the downloaded file
 
         //TODO: Refactoring
-        if (!curExt.equals(downloadExt)) {
+        if (!curExt.equals(downloadExt)) {  // convert downloaded file to the file with the current extension if these extensions aren't equal
             try {
-                String newFileUri = serviceConverter.getConvertedUri(downloadUri, downloadExt, curExt, serviceConverter.generateRevisionId(downloadUri), null, false);  // convert file and get url to a new file
+                String newFileUri = serviceConverter.getConvertedUri(downloadUri, downloadExt, curExt, serviceConverter.generateRevisionId(downloadUri), null, false, null);  // convert a file and get URL to a new file
                 if (newFileUri.isEmpty()) {
                     newFileName = documentManager
                             .getCorrectName(fileUtility.getFileNameWithoutExtension(fileName) + downloadExt);  // get the correct file name if it already exists
@@ -109,43 +111,46 @@ public class DefaultCallbackManager implements CallbackManager {
             }
         }
 
-        String storagePath = storagePathBuilder.getFileLocation(newFileName);
-        Path histDir = Paths.get(storagePathBuilder.getHistoryDir(storagePath));
-        storageMutator.createDirectory(histDir);
+        String storagePath = storagePathBuilder.getFileLocation(newFileName);  // get the path to a new file
+        Path lastVersion = Paths.get(storagePathBuilder.getFileLocation(fileName));  // get the path to the last file version
 
-        String versionDir = documentManager.versionDir(histDir.toAbsolutePath().toString(),
-                storagePathBuilder.getFileVersion(histDir.toAbsolutePath().toString(), false), true);
+        if (lastVersion.toFile().exists()) {  // if the last file version exists
+            Path histDir = Paths.get(storagePathBuilder.getHistoryDir(storagePath));  // get the history directory
+            storageMutator.createDirectory(histDir);  // and create it
 
-        Path ver = Paths.get(versionDir);
-        Path lastVersion = Paths.get(storagePathBuilder.getFileLocation(fileName));
-        Path toSave = Paths.get(storagePath);
+            String versionDir = documentManager.versionDir(histDir.toAbsolutePath().toString(),  // get the file version directory
+                    storagePathBuilder.getFileVersion(histDir.toAbsolutePath().toString(), false), true);
 
-        storageMutator.createDirectory(ver);
-        storageMutator.moveFile(lastVersion,  Paths.get(versionDir + File.separator + "prev" + curExt));
+            Path ver = Paths.get(versionDir);
+            Path toSave = Paths.get(storagePath);
 
-        downloadToFile(downloadUri, toSave);
-        downloadToFile(changesUri, Path.of(versionDir + File.separator + "diff.zip"));
+            storageMutator.createDirectory(ver);  // create the file version directory
+            storageMutator.moveFile(lastVersion, Paths.get(versionDir + File.separator + "prev" + curExt));  // move the last file version to the file version directory with the "prev" postfix
 
-        JSONObject jsonChanges = new JSONObject();
-        jsonChanges.put("changes", body.getHistory().getChanges());
-        jsonChanges.put("serverVersion", body.getHistory().getServerVersion());
-        String history = objectMapper.writeValueAsString(jsonChanges);
+            downloadToFile(downloadUri, toSave);  // save file to the storage path
+            downloadToFile(changesUri, Path.of(versionDir + File.separator + "diff.zip"));  // save file changes to the diff.zip archive
 
-        if(history==null && body.getHistory()!=null){
-            history = objectMapper.writeValueAsString(body.getHistory());
+            JSONObject jsonChanges = new JSONObject();  // create a json object for document changes
+            jsonChanges.put("changes", body.getHistory().getChanges());  // put the changes to the json object
+            jsonChanges.put("serverVersion", body.getHistory().getServerVersion());  // put the server version to the json object
+            String history = objectMapper.writeValueAsString(jsonChanges);
+
+            if (history == null && body.getHistory() != null) {
+                history = objectMapper.writeValueAsString(body.getHistory());
+            }
+
+            if (history != null && !history.isEmpty()) {
+                storageMutator.writeToFile(versionDir + File.separator + "changes.json", history);  // write the history changes to the changes.json file
+            }
+
+            storageMutator.writeToFile(versionDir + File.separator + "key.txt", key);  // write the key value to the key.txt file
+            storageMutator.deleteFile(storagePathBuilder.getForcesavePath(newFileName, false));  // get the path to the forcesaved file version and remove it
         }
-
-        if (history != null && !history.isEmpty()) {
-            storageMutator.writeToFile(versionDir + File.separator + "changes.json", history);
-        }
-
-        storageMutator.writeToFile(versionDir + File.separator + "key.txt", key);
-        storageMutator.deleteFile(storagePathBuilder.getForcesavePath(newFileName, false));
     }
 
     //TODO: Replace (String method) with (Enum method)
     @SneakyThrows
-    public void commandRequest(String method, String key) {
+    public void commandRequest(String method, String key) {  // create a command request
         String DocumentCommandUrl = docserviceUrlSite + docserviceUrlCommand;
 
         URL url = new URL(DocumentCommandUrl);
@@ -156,14 +161,14 @@ public class DefaultCallbackManager implements CallbackManager {
         params.put("key", key);
 
         String headerToken;
-        if (jwtManager.tokenEnabled())
+        if (jwtManager.tokenEnabled())  // check if a secret key to generate token exists or not
         {
             Map<String, Object> payloadMap = new HashMap<>();
             payloadMap.put("payload", params);
-            headerToken = jwtManager.createToken(payloadMap);
-            connection.setRequestProperty(documentJwtHeader.equals("") ? "Authorization" : documentJwtHeader, "Bearer " + headerToken);
+            headerToken = jwtManager.createToken(payloadMap);  // encode a payload object into a header token
+            connection.setRequestProperty(documentJwtHeader.equals("") ? "Authorization" : documentJwtHeader, "Bearer " + headerToken);  // add a header Authorization with a header token and Authorization prefix in it
 
-            String token = jwtManager.createToken(params);
+            String token = jwtManager.createToken(params);  // encode a payload object into a body token
             params.put("token", token);
         }
 
@@ -171,36 +176,43 @@ public class DefaultCallbackManager implements CallbackManager {
 
         byte[] bodyByte = bodyString.getBytes(StandardCharsets.UTF_8);
 
-        connection.setRequestMethod("POST");
-        connection.setRequestProperty("Content-Type", "application/json; charset=UTF-8");
-        connection.setDoOutput(true);
+        connection.setRequestMethod("POST");  // set the request method
+        connection.setRequestProperty("Content-Type", "application/json; charset=UTF-8");  // set the Content-Type header
+        connection.setDoOutput(true);  // set the doOutput field to true
         connection.connect();
 
         try (OutputStream os = connection.getOutputStream()) {
-            os.write(bodyByte);
+            os.write(bodyByte);  // write bytes to the output stream
         }
 
-        InputStream stream = connection.getInputStream();
+        InputStream stream = connection.getInputStream();  // get input stream
 
         if (stream == null) throw new RuntimeException("Could not get an answer");
 
-        String jsonString = serviceConverter.convertStreamToString(stream);
+        String jsonString = serviceConverter.convertStreamToString(stream);  // convert stream to json string
         connection.disconnect();
 
-        JSONObject response = serviceConverter.convertStringToJSON(jsonString);
+        JSONObject response = serviceConverter.convertStringToJSON(jsonString);  // convert json string to json object
         //TODO: Add errors ENUM
-        if (!response.get("error").toString().equals("0")){
-            throw new RuntimeException(response.toJSONString());
+        String responseCode = response.get("error").toString();
+        switch(responseCode) {
+            case "0":
+            case "4": {
+                break;
+            }
+            default: {
+                throw new RuntimeException(response.toJSONString());
+            }
         }
     }
 
     @SneakyThrows
-    public void processForceSave(Track body, String fileName) {
+    public void processForceSave(Track body, String fileName) {  // file force saving process
 
         String downloadUri = body.getUrl();
 
-        String curExt = fileUtility.getFileExtension(fileName);
-        String downloadExt = fileUtility.getFileExtension(downloadUri);
+        String curExt = fileUtility.getFileExtension(fileName);  // get current file extension
+        String downloadExt = fileUtility.getFileExtension(downloadUri);  // get an extension of the downloaded file
         Boolean newFileName = false;
 
         // convert downloaded file to the file with the current extension if these extensions aren't equal
@@ -208,7 +220,7 @@ public class DefaultCallbackManager implements CallbackManager {
         if (!curExt.equals(downloadExt)) {
             try {
                 String newFileUri = serviceConverter.getConvertedUri(downloadUri, downloadExt,
-                        curExt, serviceConverter.generateRevisionId(downloadUri), null, false);  // convert file and get url to a new file
+                        curExt, serviceConverter.generateRevisionId(downloadUri), null, false, null);  // convert file and get URL to a new file
                 if (newFileUri.isEmpty()) {
                     newFileName = true;
                 } else {
@@ -226,18 +238,18 @@ public class DefaultCallbackManager implements CallbackManager {
         boolean isSubmitForm = body.getForcesavetype().toString().equals("3");
 
         //TODO: Extract function
-        if (isSubmitForm) {
+        if (isSubmitForm) {  // if the form is submitted
             if (newFileName){
                 fileName = documentManager
                         .getCorrectName(fileUtility.getFileNameWithoutExtension(fileName) + "-form" + downloadExt);  // get the correct file name if it already exists
             } else {
                 fileName = documentManager.getCorrectName(fileUtility.getFileNameWithoutExtension(fileName) + "-form" + curExt);
             }
-            forcesavePath = storagePathBuilder.getFileLocation(fileName);
+            forcesavePath = storagePathBuilder.getFileLocation(fileName);  // create forcesave path if it doesn't exist
             List<Action> actions =  body.getActions();
             Action action = actions.get(0);
-            String user = action.getUserid();
-            storageMutator.createMeta(fileName, user, "Filling Form");
+            String user = action.getUserid();  // get the user ID
+            storageMutator.createMeta(fileName, user, "Filling Form");  // create meta data for the forcesaved file
         } else {
             if (newFileName){
                 fileName = documentManager.getCorrectName(fileUtility.getFileNameWithoutExtension(fileName) + downloadExt);
